Use a relational persistance layer and other things
MongoDB didn't lend itself well to the way we specified the HFCu
interface. This adds classes and code to use the javax.persistance.*
annotations to use a relational database layer. Currently it is an
hibernate (in memory) database, but switching to another (like mysql)
database shouldn't be hard. The downside of hibernate is that data doesn't
persist between service restarts.

During the hackathon, I also worked on implementing the other endpoints,
so this is a complete prototype. It is still missing all kinds of access
control and checks (e.g. not null checks for the controller and much
more).